Restructure internal tests
